voice_wake_word/requirements.txt
# 语音唤醒系统依赖项

# 核心依赖
pvporcupine>=3.0.0       # Porcupine语音唤醒引擎
pyaudio>=0.2.13          # 音频输入输出库

# 可选依赖
numpy>=1.21.0            # 数值计算库
psutil>=5.9.0            # 系统资源监控
soundfile>=0.10.3        # 音频文件处理
matplotlib>=3.5.0        # 可视化工具（可选）
python-dotenv>=1.0.0      # 环境变量加载（可选，便于读取 .env）

# 开发依赖
pytest>=7.0.0            # 测试框架
flake8>=4.0.0            # 代码检查
black>=22.0.0            # 代码格式化